moved IO helper functions to IOUtils.py
