--- tests/EOS/__init__.py
@@ -1,1 +1,0 @@
-from .timer import StopWatch, Measure, Profiling
